Created a new massively improved rendering system
Created a significantly more advanced rendering system capable of
dealing with any amount of arbitrary elements, placing them next to
eachother and on the screen with (almost) minimal mallocing. Added a
qw_ in front of all structs and enums defined by me so there is
effectively zero chance of ever running into a naming conflict with any
headers I might make use of in the future.